Automate establishment of OAuth credentials.
The progression is not yet quite right for the initial setup - the script
needs to be run twice to do the process, with redundant credential
stashing. I will resolve that next, but most of the code for the process is
now there and working, warranting a commit.

The authentication setup includes collecting a device id, selected by the
user. The device ID is stashed in a file in the script directory named
`.device_id`.